completed offer_app

+++ offer_app/models.py
@@ -40,7 +40,7 @@ class OfferDetail(models.Model):
     title = models.CharField(max_length=100, blank=False, default='')
     revisions = models.IntegerField(default=0)
     delivery_time_in_days = models.IntegerField(default=0)
-    price = models.DecimalField(max_digits=10, decimal_places=0, default=0)
+    price = models.IntegerField(default=0)
     features = models.JSONField(default=list)
     offer_type = models.CharField(max_length=50, default='Basic')
